move all version-related functions into version.py
(cherry picked from commit 95753e4513736abcf48aa8ef9c6900818935019e)
